Version 1: New API (#40)
Public API revamp, provide a way of multiple wrappers be stacked.

```swift
@ObservableValue
@RemoteValue(.backend(.getUserName))
@Persistent(.secure("com.user.name"))
var str = "str-default"
```


Also changing the mutability:

```swift
@SwiftUIBind(
    \Storage.$str,
     mutate: UpdateStr.self
) var str
```


So `@Mutable` wrapper is removed, instead, decisions are the only way to
mutate the state.

`UpdateStr` is a special type of decision, that takes a `newValue` as a
parameter:
```swift 
struct UpdateStr: ValueDecision {
    var newValue: String

    func mutate(_ env: Decide.DecisionEnvironment) {
        env[\.Storage.$str] = newValue
    }
}
```


this way the mutability is expressed explicitly and gives ability to
combine other decisions and side effects on value updates.



**Verbosity:** right now this version requires quite a sophisticated
code to be written, e.g. `SwiftUIBind`, later the plan is to utilise
swift macros to write this code for the user.